requirements.txt
# SentriKat Dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
APScheduler==3.10.4
requests==2.32.4
python-dotenv==1.0.0
gunicorn==22.0.0
ldap3==2.9.1
reportlab==4.0.7
email-validator==2.1.0
cryptography==46.0.5
qrcode==7.4.2
Pillow==12.1.1
PyYAML==6.0.1

# Database
psycopg2-binary==2.9.9

# SAML SSO (optional - requires system packages: libxml2-dev libxmlsec1-dev)
python3-saml==1.16.0

# OAuth/OIDC SSO (optional)
Authlib==1.4.1

# WebAuthn/FIDO2 hardware security key support (optional)
webauthn==2.7.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
